requirements.txt
flask==2.3.3
selenium==4.12.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
requests==2.31.0 
gunicorn==21.2.0
gspread==5.1.1
oauth2client==4.1.3
playwright==1.40.0 